# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Loginex TMS API
  description: |
    API для диспетчерской системы Loginex TMS.
    - Auth: серверные сессии в БД. В cookie — подписанный HMAC-SHA256 токен с jti = id строки Session
      (loginex_staff_session для сотрудников, loginex_driver_session для водителей).
      Отзыв строки Session немедленно закрывает доступ.
    - Роли: admin | logist (кабинет), driver (мобильное приложение).
      driverId и принадлежность организации берутся из сессии, а не из тела запроса.
    - Rate limiting: 5 попыток / 15 мин для auth
    - CSRF: double-submit cookie loginex_csrf + header x-csrf-token
    - Validation: Zod, 400 с details.issues
  version: 1.0.0
  contact:
    name: Loginex TMS
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://loginex.example.com
    description: Production
...
paths:
...
  /api/vehicles:
    get:
      tags: [Vehicles]
      summary: List vehicles
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Vehicles
    post:
      tags: [Vehicles]
      summary: Create vehicle
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicle'
      responses:
        '200':
          description: Created
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /api/chat:
    get:
      tags: [Chat]
      summary: List chat messages
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Messages
    post:
      tags: [Chat]
      summary: Send message
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  maxLength: 2000
                recipientId:
                  type: string
                type:
                  type: string
                isImportant:
                  type: boolean
      responses:
        '200':
          description: Sent
        '400':
          $ref: '#/components/responses/ValidationError'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: loginex_token
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    CreateVehicle:
      type: object
      required: [plate, type, capacity]
      properties:
        plate:
          type: string
          maxLength: 20
        type:
          type: string
        capacity:
          type: integer
          minimum: 1
        year:
          type: integer
        volume:
          type: number
        length:
          type: number
        width:
          type: number
        height:
          type: number
...
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Validation failed
              details:
                type: object
                properties:
                  issues:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: array
                          items:
                            type: string
                        message:
                          type: string
...